Gestion de cinema
-ajout de tout les fonctionnalites avec alea

+++ src/main/java/repository/CategoriePersonneRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CategoriePersonneRepository extends JpaRepository<CategoriePersonne, Long> {
+public interface CategoriePersonneRepository extends JpaRepository<CategoriePersonne, Integer> {
 }
