+++ src/main/java/repository/SalleRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SalleRepository extends JpaRepository<Salle, Long> {
+public interface SalleRepository extends JpaRepository<Salle, Integer> {
 }
