+++ src/main/java/repository/TypePlaceRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TypePlaceRepository extends JpaRepository<TypePlace, Long> {
+public interface TypePlaceRepository extends JpaRepository<TypePlace, Integer> {
 }
